requirements.txt
ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
opencv-python>=4.5.0
pillow>=8.3.0
numpy>=1.21.0
matplotlib>=3.4.0
albumentations>=1.3.0
roboflow>=1.2.11
python-dateutil>=2.9.0